pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "FastAPI layered architecture boilerplate"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "pydantic>=2.5.0",
  "pydantic-settings>=2.2.0",
  "sqlalchemy>=2.0.0",
  "asyncpg>=0.29.0",
  "alembic>=1.13.0",
  "PyJWT>=2.8.0",
  "passlib[bcrypt]>=1.7.4",
  "prometheus-client>=0.20.0",
  "prometheus-fastapi-instrumentator>=6.1.0",
  "redis>=5.0.0",
  "aio-pika>=9.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.26.0",
  "ruff>=0.3.0",
  "black>=24.2.0",
  "aiosqlite>=0.20.0",
  "fakeredis>=2.22.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
